requirements.txt
numpy
pandas
scikit-learn
imbalanced-learn
joblib
tensorflow-metal